requirements.txt
Pillow>=10.0.1, <11
Wand>=0.6.11, <0.7